refactor(download) Extract downloadBlob function
