+++ src/lib/utils/download.ts
@@ -1,12 +1,8 @@
+import { downloadBlob } from './download/index.js'
+
 export async function downloadFile(src: string, filename: string) {
   const response = await fetch(src)
   const blob = await response.blob()
-  const url = URL.createObjectURL(blob)
-  const link = document.createElement('a')
 
-  link.href = url
-  link.download = filename
-  link.click()
-
-  URL.revokeObjectURL(url)
+  downloadBlob(blob, filename)
 }
